gdk: Make gdk_drag_begin take doubles
This is a better fit to the way we treat coordinates
everywhere else.

+++ gdk/broadway/gdkdnd-broadway.c
@@ -87,8 +87,8 @@ _gdk_broadway_surface_drag_begin (GdkSurface         *surface,
                                   GdkDevice          *device,
                                   GdkContentProvider *content,
                                   GdkDragAction       actions,
-                                  gint                dx,
-                                  gint                dy)
+                                  double              dx,
+                                  double              dy)
 {
   GdkDrag *new_context;
