+++ gdk/wayland/gdkdrag-wayland.c
@@ -352,8 +352,8 @@ _gdk_wayland_surface_drag_begin (GdkSurface         *surface,
                                  GdkDevice          *device,
                                  GdkContentProvider *content,
                                  GdkDragAction       actions,
-                                 gint                dx,
-                                 gint                dy)
+                                 double              dx,
+                                 double              dy)
 {
   GdkWaylandDrag *drag_wayland;
   GdkDrag *drag;
